drm: add virtual CRTC manager (VCRTCM) from Bell Labs
VCRTCM allows easy creation of additional (virtual) CRTCs to the
GPU and linking them to external device for display transmission
(to remote display) with or without compression. The augmented GPU
appears to the rest of the system like it has additional
connectors/encoders and additional CRTCs which are recognized by the
windowing system and used as "equal citizens" as other real (a.k.a.
physical) CRTCs.

Signed-off-by: Ilija Hadzic <ihadzic@research.bell-labs.com>