drm/radeon/kms: new parameter for vblank emulation
the right way to do vblank emulation in push mode is to
emit a fence after blit_copy request and then when the fence
is signaled issue vblank emulation request; the problem is
that there is no good mechanism to do real-time notification
when the fence is signaled (they are all sort-of "garbage
collected" when some other process requests to do something
with, possibly different, fence). We need a fence-signaled
interrupt, which may be doable using SCRATCH_INT interrupt
(which is currently unused in radeon driver), but until we
figure out how to put it into use, we have a less fortunate
solution: as soon as copy request is issued, singal vblank
and let the application proceed and submit new rendering
commands. They will anyway happen after the copy due
to FIFO nature of the ring, so we are safe when it comes
to frame tearing. The only "fuzzy" part is that we just
"trust" the GPU that copy will comnplete within the frame time.

Signed-off-by: Ilija Hadzic <ihadzic@research.bell-labs.com>