udlctd: Shortened often used variable name
Shortened the variable name udlctd_vcrtcm_hal_descriptor
to uvhd wherever it is used. The long form is still used
in the actual name of the structure, and as the variable
name when it is contained in another structure.

This was done to make it easier to keep lines under 80
characters.

Also cleaned up any noticed line wraps that became unnecessary
as a result of this change.

Signed-off-by: William Katsak <william.katsak@alcatel-lucent.com>