udlctd: Major cleanup + fixed pitch + minor issues
Reworked allocation of scratch memory. We now allocate
the backing_buffer, hline_16, and hline_8 as whole pages
and map them manually using vm_map_ram(), rather than
using vmalloc() for the entire buffers.

Eliminated pull mode. This mode has long been deprecated
in udlctd, and was only there because it was the first
proof of concept during development. This modification
allowed for the elimination of the local front framebuffer
because the push buffers fill this role.

The removal of the front framebuffer created a problem for
the current technique used to generate the "bluescreen" on
initial driver load and mode switch. The old method of
doing this was to simply fill the local front buffer with
the blue pixels. The new method instead kmallocs enough
memory for a single hline, fills it with the color pixels,
and retransmits the line y-resolution times. Also moved
the invocation of this function to inside set_video_mode.

This patch also fixes the distorted output that was seen
at resolutions where the framebuffer pitch did not match
the x-resolution.

Signed-off-by: William Katsak <william.katsak@alcatel-lucent.com>