udlctd: Handle being set to an invalid resolution
This fixes the crash that occured if a program running
on a VCRTC attached to a DisplayLink device changed the
resolution to one that is unsupported by the DL/monitor
combination.

For now, the solution is to allow VCRTCM think that the
mode switch was successful, but display an error screen
on the physical device and disallow any transmission
over USB by forcing FPS to be zero for as long as the
VCRTC is in the invalid mode.

Signed-off-by: William Katsak <william.katsak@alcatel-lucent.com>