drm/radeon/kms: handle pflip on virtual crtc with no hal
If a virtual crtc has nothing attached to it and a page
flip ioctl is called, there will be no vblanks to handle
the flip. In this case all of flip handling must be done
in ioctl, otherwise the process flipping on unattached
CRTC will hang up. Since there are no vblanks, we have to
make up the sequence number and the timestamp, but
we have no other choice.

Signed-off-by: Ilija Hadzic <ihadzic@research.bell-labs.com>