vcrtcm: add gpu method to check connected status
Add new method called 'connected' that allows the GPU to check
whether a CTD device is connected to the display it is driving.
It only makes sense for CTD devices that drive actual display.
Such devices would want to implement a 'connected' backend
and query the display they are driving in it. Devices that
emulate displays in software do not need to implement this method
and VCRTCM layer will consider that them always connected.

Signed-off-by: Ilija Hadzic <ihadzic@research.bell-labs.com>