drm/radeon/kms: add vcrtc-always-connected param
Add new module parameter to force virtual CRTC to be
always conncted. If it's always connected, GPU driver
should blindly return the connected status and make up
some commonly used modes. Otherwise it should
consider all CRTCs that are not attached as "not connected"
and if they are attached, ask the VCRTCM module for status
information.

This patch only introduces the parameter, but it
does not use it (yet).

Signed-off-by: Ilija Hadzic <ihadzic@research.bell-labs.com>