vcrtcm: add get_modes and check_mode methods
Add two new methods (get_modes and check_mode) that a GPU
can use to check with the attached CTD which modes are available
and if a particular mode is valid. CTDs that drive real displays
should implement these methods to query the capabilities of
the displays they are driving.

Signed-off-by: Ilija Hadzic <ihadzic@research.bell-labs.com>