drm/radeon/kms: use vcrtcm to query virtual crtc modes and status
VCRTCM now provides ability for a GPU to query the attached CTD
device for supported modes and status. Hook up relevant
CRTC functions to this backend. Also start using conn_virt_crtc
paramter: when 1 use old semantics in which a virtual crtc is
always connected; when 0, use new semantics in which virtual
crtc disconnected if there is no CTD attached, while it
asks the VCRTCM if there is something attached. Follow the
same logic for querying modes: if there is nothing attached, no
modes are vaild; if something is attached, ask VCRTCM.

Signed-off-by: Ilija Hadzic <ihadzic@research.bell-labs.com>