udlctd: fix a bug in allocation of minor numbers
Fixed a bug that appeared in the case where more than one
DisplayLink dongles are attached, then a lower-minor-numbered
one is removed and reconnected.

Before, the reconnected dongle would be incorrectly be assigned
a minor that was already assigned to another device, causing
problems when trying to attach the device to a CRTC. The root cause
of this was an incorrect algorithm for assigning a minor; we simply
set each new device to num_minors, then incremented that value by
one.

Now we use a better algorithm, and explictly try to reassign free
minors that are lower than the current maximum minor. This algorithm
uses a 64-bit value to bitwise keep track of used minors, so we also set
UDLCTD_MAX_DEVICES to 64 and refuse to register a device when there
are already that many present.

Signed-off-by: William Katsak <william.katsak@alcatel-lucent.com>