drm/radeon/kms: make viewport_y zero in push mode
In push mode, GPU carves out only the visible section of
the frame buffer in vertical dimension and pushes the whole
pitch in horizontal dimension (for now).

In pull mode, GPU does not push anything, but instead just
passes the pointer and the dimensions information to the
CTD driver and lets the CTD driver find the pixels it needs.

viewport_y must be consistent with what is pushed, otherwise
CTD driver may look at the wrong pixels in the driver (CTD
driver is still supposed to honor viewport_y and viewport_y).

To fix this, GPU now looks at the hw_prop field to find
out in which mode the CTD is. If it's in the push mode,
then it sets viewport_y to zero, otherwise (for pull mode)
it uses y from radeon_crtc structure as is.

Signed-off-by: Ilija Hadzic <ihadzic@research.bell-labs.com>