udlctd: changed how current mode is stored
Previously, we kept a pointer inside struct udlctd_info to
a struct udlctd_video_mode, which was kept in a list of
valid modes.

However, if this list was repopulated (if the modes were
redetected, for example), that pointer could be made invalid.

Instead, keep the information directly inside the struct
udlctd_info and copy the mode struct into that memory upon
mode change.

Signed-off-by: William Katsak <william.katsak@alcatel-lucent.com>