drm/radeon/kms: rework fb setup for virtual crtc
radeon_vcrtcm_set_fb function does not necessarily set
the frame buffer associated with the crtc. When called
with atomic=1, it uses the frame buffer supplied in
the function argument. This patches reworks this function
to make sure that the right frame buffer is passed
to vcrtcm_set_fb and also moves some code around to
make the whole thing look better with less repeated
code and redundant data.

Signed-off-by: Ilija Hadzic <ihadzic@research.bell-labs.com>