drm/radeon/kms: fix set_fb when atomic is 1
When radeon_vcrtcm_set_fb is called with atomic=1, the
frame buffer is not a CRTC's buffer, but it is supplied
by the caller. This case is typically used when console is
taken over by the kernel debugger. In this case, crtc->fb
is no longer a frame buffer from which a virtual CRTC should
read, but instead a temporary one.

This patch adds a new field to radeon_crtc to track the
frame buffer from which GPU pushes pixels to VCRTCM
and makes use of it in relevant functions (push, page_flip
and sef_fb). This ensures that the right buffer is always
pushed to VCRTCM.

Signed-off-by: Ilija Hadzic <ihadzic@research.bell-labs.com>