udlctd: fix scripted attach/set_fps failure
Fix a bug where if attach() and set_fps() are executed too close
together, e.g. in a script, set_fps() would fail to succeed.

This bug was a result of the technique being used to prevent
transmission upon an invalid mode set.

Previously, set_fps() checked to see if an "xmit_allowed"
flag was set before it started the vblank thread. This flag
was set upon a valid mode set, and cleared upon a bad mode set.

Since directly after attach(), there is a short delay before the
modes are properly queried and the initial modeset occurs, a very
quick subsequent call to set_fps() would fail to start the vblank
thread, resulting in the need to issue another set_fps() call to
get display output.

This patch reworks the behavior of set_fps() to always start
the vblank thread when called, and modifies udlctd_do_xmit_fb_push()
to not transmit and fake out vblanks as long as the xmit_allowed
flag is unset, indicating that the screen is in a mode that is invalid
for the DisplayLink dongle. Upon a valid mode set, the xmit_allowed
flag is set and normal transmission commences, without the need for
any additional calls to set_fps().

Signed-off-by: William Katsak <william.katsak@alcatel-lucent.com>