vcrtcm: add new disable GPU method + CTD callback
Add a new function for the GPU to call when a crtc
is disabled. Additionally, add the function pointer
for a corresponding CTD callback.

This is necessary because when a vcrtc is disabled,
e.g. by xrandr, the framebuffer associated with it
can be resized.

If CTDs are not made aware of this change, they can
end up in a situation where the CTD driver is still
trying to get pixels from a framebuffer that no longer
exists.

Signed-off-by: William Katsak <william.katsak@alcatel-lucent.com>