drm: hold mutex in critical sections of render-node code
Critical sections are parts of the code where we claim or
release resources (we don't want two render-node create or
remove ioctl called in the context of different processes
to claim part of requested resources because of the race).
Another critical section is manipulating the render node
list. We can use dev->struct_mutex for both.

Signed-off-by: Ilija Hadzic <ihadzic@research.bell-labs.com>