drm: fix encoders list returned by drm_mode_getconnector
Restrict the list of (possible/available) encoders
returned by drm_mode_getconnector to only those that actually
belong to the mode group.

User space can create a render node with a connector that can be
used with multiple encoders but include only a subset of
possible encoders in the mode group. This patch ensures that
encoders that were not included at node creation time are not
returned to the user space when drm_mode_getconnector
is called.

Signed-off-by: Ilija Hadzic <ihadzic@research.bell-labs.com>