vcrtcm: rename interface-function structs
vcrtcm currently defines the structs "vcrtcm_funcs" and "vcrtcm_gpu_callbacks".
These names are confusing because the functions in the former struct are not
actually vcrtcm functions, they are the functions comprising the interface
to the pcon.  This commit changes the struct names to "vcrtcm_pcon_funcs"
and "vcrtcm_gpu_funcs".  This commit also fixes all code that breaks as a
result of these changes.

Signed-off-by: Martin Carroll <martin.carroll@alcatel-lucent.com>