vcrtcm: rename HAL to PCON
The vcrtcm's "HAL" is not actually a HAL.  It is simply a
functional interface to a PCON.  Whether or not the PCON also
happens to be a driver for a piece of hardware is hidden from,
and irrelevant to, vcrtcm.  To reflect this fact, all vcrtcm
identifiers and comments mentioning HAL have been renamed to
mention PCON instead.

Signed-off-by: Martin Carroll <martin.carroll@alcatel-lucent.com>