vcrtcm: lexically separate two sides of vcrtcm interface
vcrtcm actually exposes two separate interfacess.
One of the interfaces comprises the functions that GPU drivers
call, the other interface comprises the functions that PCONs call.
The current vcrtcm function-naming scheme does not indicate which
of the two interfaces a given function belongs to.  This commit
lexically separates the two interfaces.

Signed-off-by: Martin Carroll <martin.carroll@alcatel-lucent.com>