vcrtcm: rename xmit_fb to dirty_fb
The xmit_fb() function that PCONs are required to implement is
misnamed.  The implementation of this function is not required
to transmit the framebuffer; indeed, the PCON is not required
ever to transmit *any* framebuffers.  In fact, a call to this
function is merely an indication to the PCON that the current
framebuffer has been modified.  The PCON can handle that modification
however it likes.

Signed-off-by: Martin Carroll <martin.carroll@alcatel-lucent.com>