vcrtcm: change lexical separation of vcrtcm's two interfaces
This commit changes the prefixes that are used to lexically
separate vcrtcm's two interfaces.  Now the gpu-side functions are
prefixed with vcrtcm_g_, and the PCON-side functions are prefixed
with vcrtcm_p_.  This scheme makes it easier to grep for these
functions in source code, and also eliminates the semantic
confusion that could result with the previous scheme.

Signed-off-by: Martin Carroll <martin.carroll@alcatel-lucent.com>