vcrtcm: rename p_gpu_sync function p_wait_fb
The GPU side of the vcrtcm API has a function "g_wait_fb" to enable
the GPU to wait for the PCON to finish with the frame.  The PCON
side of the vcrtcm API had (before this commit) a function "p_gpu_sync"
to enable the PCON to wait for the PCON to finish with the frame.
This commit makes the names analogous, to match the fact that the
functionality is analogous.

Signed-off-by: Martin Carroll <martin.carroll@alcatel-lucent.com>