drm/radeon/kms: fix vcrtcm crash due to blit rework
Patch f237750f007412eb5e1baafe3e32857b35fbc6ee reworked
blit copy functions quite a bit and vcrtcm's use of copy
must be reconciled with it.

Fences are now mandatory in blit copy functions (can't pass
NULL) because the blit-related bo is now dynamically
allocated (so a fence is needed to tell when to free it).

So the caller now must always create the fence even if
it does not care about waiting for copy to finish. The
non-caring caller must unref the fence. The radeon_sa_bo_free
structure will raise another reference to the fence
(so it won't go away) if it needs to wait on it before
freeing the bo.

Signed-off-by: Ilija Hadzic <ihadzic@research.bell-labs.com>