drm/radeon/kms: initialize vcrtcm copy fences to NULL
Reworked blit-copy code now creates its own fence and requires
a pointer to the fence pointer be passed to it, so that it can
populate it upon return.

The caller's ponter must be initialized to either NULL or a
valid fence, otherwise radeon_fence_need_sync in ***_prepare_copy
will misbehave.

Signed-off-by: Ilija Hadzic <ihadzic@research.bell-labs.com>