vcrtcm: allow detach PIM function to fail
Some PIMs can fail on detach and the failure should be communicated
back to the user space application trying to detach. Specific example
is v4l2pim that attaches to a CRTC and creates a new v4l2 device
that an application can open. If a detach request comes in while
the video device is open, it is very hard (almost impossible) to
properly cleanup the state. In this situation, detach should fail
(typically returning -EBUSY).

To support this use case, modify detach method to allow failure
and pass the error code. PIMs that cannot fail on detach, should
simply return 0.

Signed-off-by: Ilija Hadzic <ihadzic@research.bell-labs.com>