vcrtcm: add functions to export push buffers to PRIME
In the new scheme of things, VCRTCM will no longer ask the GPU
to create a push buffer for it (i.e., the pb_alloc call in GPU
will no longer be used) but instead PCON will create its backing
store and map it such that it can access its data (which is
PCON-dependent). Then PCON would call a VCRTCM function to export
the buffer using DMA-BUF sharing infrastructure, followed by
calling DRM/PRIME function to turn the buffer into a GEM object
to which the GPU's push function can copy the pixels.

This patch adds two new GPU methods (and a bunch of supporting code)
that PCONs will call to register and unregister the push buffers with
PRIME.

Signed-off-by: Ilija Hadzic <ihadzic@research.bell-labs.com>