vcrtcm: add memory manipulation utilities
All PIMs do mostly the same thing when it comes to allocating
memory. They either want a page, a group of pages, a chunk of
continuous memory in vmalloc region or a chunk of continuous
memory for kernel's use (kmalloc).

While doing that they also use their own counters that come
handy when debugging memory leaks. Rather than repeating all
this code in each PIM, we factor out common functions into
VCRTCM and export them to PIMs.

This patch is a prep for eliminating a bunch of repeated
code from PIMs.

Signed-off-by: Ilija Hadzic <ihadzic@research.bell-labs.com>