vcrtcm: go back to using printk for debug messages
"pr_debug ..." does not seem to be of equivalent log level
and facility as "printk(KERN_DEBUG ...", so on some machines
debug messages from PIMs and VCRTCM get filtered out before
they make it to the log file.

Until we figure out the proper way to fix it, let's fix it
the "wrong" way by reverting to using printk for debug
messages.

Signed-off-by: Ilija Hadzic <ihadzic@research.bell-labs.com>