pimmgr: support properties on pim interface
Add a function get_properties() to the pim interface. When this function
is called by pimmgr (if the PIM implements it), the PIM will populate the
struct pimmgr_pcon_properties that is passed as a parameter.

Properties are exposed via sysfs.

Initially, we support two new properties: fps and attached:

- fps is an integer value for the PCON's current frame rate (frames per
second).

- attached is an integer (boolean) value to determine if the PCON is
currently attached to a VCRTC.

Signed-off-by: William Katsak <william.katsak@alcatel-lucent.com>