pimmgr: change pconids to low sequential integers
Previously, pconids were a large unsigned integer created by
combining pimids and and local_pconids with a bitshift sequence.

This resulted in large, opaque integers near 2^31. It has since
been decided that small, sequential numbers similar to file
descriptors would be a better choice.

With this patch, pconids are now simply the offset into a list
of mappings (pconid -> pimid/local_pconid).

This patch changes the pconid type from uint32_t to int, creates
a the lookup table and helper functions, and cleans up some
variations in the names of pconid/pimid/local_pconid variables.

This patch also changes the pimmgr ioctl error scheme to be
more consistent with the errno standard and adds an additional
error code (out of pconids).

Signed-off-by: William Katsak <william.katsak@alcatel-lucent.com>