pimmgr: change signatures of pim_funcs
This commit changes the signature of the destroy() and get_properties()
functions of pim_funcs.  The following series of commits will make the
corresponding fixes in each of the pims.  Because this change is broken
up over several commits, the tree as of this commit does not build
without warnings, and the pims do not work correctly.

v2: fix formatting

Signed-off-by: Martin Carroll <martin.carroll@alcatel-lucent.com>
Signed-off-by: Ilija Hadzic <ihadzic@research.bell-labs.com>