vcrtcm: eliminate one of the pcon properties structs
There were two pcon-properties structs, one of which was used
to hold a single static property of the pcon.  All other static
properties of the pcon are stored in the pcon_info struct.  This
commit eliminates the "rump" static-properties struct and moves
the one property therein to the pcon_info struct, where it belongs.
Note that this commit breaks the build of the pims; successive
commits will fix the pims.

Signed-off-by: Martin Carroll <martin.carroll@alcatel-lucent.com>