vcrtcm: add virgin flag to the push buffer
Virgin push buffer is the buffer that has the backing
store memory, it is perfectly usable, but it has never been
pushed to. Such a buffer would contain random data.

When a new buffer is created, its virgin flag is set to 1,
and when it is first pushed to, its virgin flag is set to 0.
PIMs can use this flag to avoid using garbage data (typically
in the first frame or first cursor sprite).

Signed-off-by: Ilija Hadzic <ihadzic@research.bell-labs.com>