udlpim: do not display virgin push buffer
This patch prevents displaying garbage data in first
frame or first appearance of a cursor. UDLPIM uses two
push buffers, one to read the displayed data from and
one to push the next frame data into. In the initial state
both buffers are virgin, and we push to the first buffer
followed by calling the function that copies the other push
buffer to the UDL device on USB interface. The problem
is that this other buffer is still virgin and thus contains
random data. In other words, the pipeline has not been
fully established, so the first frame or a first appearance
of the cursor can be garbage.

This patch fixes a bug that manifested itself in a momentary
appearance of a big white square when cursor was first time
entering the screen.

Signed-off-by: Ilija Hadzic <ihadzic@research.bell-labs.com>