v4l2pim: do not send a virgin push buffer to v4l2
This patch prevents sending garbage data to video 4 linux
subsystem (cursor or frame). This can happen for the first
frame or when the cursor is displayed for the first time.

After first push, pipeline is not fully established and
the back-buffer (the one that is sent to v4l2 subsystem)
is still virgin and therefore contains random data. Such a
buffer should not be sent.

Signed-off-by: Ilija Hadzic <ihadzic@research.bell-labs.com>