vcrtcm: fix races in pim registration and change unregister api
This commit fixes several race conditions that existed in the
pim registration and deregistration logic.  It also refactors
that logic.  Finally, it changes the unregister api to take a
pimid instead of a pim name.  Note that this commit breaks the
build; the following sequence of commits will fix the build.

Signed-off-by: Martin Carroll <martin.carroll@alcatel-lucent.com>