drm/radeon/kms: remove asynchronous vblank emulation
We never got it to work because fences are not signaled
in real time, but in a lazy manner (i.e. only when somebody
asks for fence status). Adding real-time fence signaling is
quite a project and will diverge the driver even more from
upstream. So just nuke the whole feature and use "speculative"
vblank emulation that we have been using all along.

Signed-off-by: Ilija Hadzic <ihadzic@research.bell-labs.com>