vcrtcm: remove counter argument from vcrtcm_kfree
The implementation of vcrtcm_kfree is eventually going to figure
out which internal counter to decrement.  This commit also
temporarily removes the FIXME comment and undoes the workaround
described in that comment.  The comment, along with a new
implementation of the workaround, will be added back in a much
later commit.  Note that this commit breaks the build; the
following sequence of commits will fix the build.

Signed-off-by: Martin Carroll <martin.carroll@alcatel-lucent.com>