udlpim: start and cancel worktasks on minor create and destroy
The two minor-related worktasks were incorrectly being started
and cancelled in the wrong places.  One problem that thereby
resulted was the fact that if a pcon was created but not attached
and then detached, one of the associated minor's worktasks never
got cancelled, even when the minor was destroyed -- hence a kernel
panic when the worktask tried to access the freed minor.  The
solution is to start the worktasks when the minor is created and
cancel them when the minor is destroyed.

Signed-off-by: Martin Carroll <martin.carroll@alcatel-lucent.com>