drm/radeon/kms: nop out wait_mem_reg for virtual crtc
VLINE_START_END command for the GPU and virtual CRTCs don't
mix very well because inside the GPU, CP and shader processors
don't know much about virtual CRTCs. The undesirable result is
a GPU hang whenever the user space sends VLINE_START_END
command referring to a virtual CRTC.

As a workaround, we nop-out the wait_mem_reg command in the
batch until we figure out what we can wait on in lieu of.

Signed-off-by: Ilija Hadzic <ihadzic@research.bell-labs.com>