vcrtcm: correctly handle invalid pimid for buffer owner
The VCRTCM_OWNER_PIM bit must be checked before checking the
VCRTCM_OWNER_VCRTCM bit, because a common pim bug is for the
pim to specify an owner of (VCRTCM_OWNER_PIM | -1), which has
both of those bits set, and we want to treat that value as
VCRTCM_OWNER_PIM.

Note that with this fix, the kernel will still build but none
of the pims will work correctly (specifically, all their
early allocations will fail).  This will be fixed in the
following sequence of commits.

Signed-off-by: Martin Carroll <martin.carroll@alcatel-lucent.com>