udlpim: remove major device number allocation
We no longer need a major nor minor device number to identify
the PIM (or PCON). Also, this PIM does not interact
with the user space directly (but indirectly through VCRTCM
and USB subsystem). So there is no reason (any more) to register
it as a character device nor allocate its major device number.

Signed-off-by: Ilija Hadzic <ihadzic@research.bell-labs.com>