vcrtcm: remove all use of pcon spin lock
The pcon locking scheme is being rearchitected.  This is the
first commit in a series of commits that implement the new scheme.
This first commit removes all use of the per-pcon spin lock.
The system after this commit still works well enough to pass
simple tests, but it temporarily has even more race conditions
than it already had.  Those race conditions will be gradually
removed over the course of the following commits.

Signed-off-by: Martin Carroll <martin.carroll@alcatel-lucent.com>