vcrtcm: protect page_flip against races
The function vcrtcm_g_page_flip is allowed to be called in atomic context,
which means that the caller is not require to have the pcon's mutex before
calling the function.  The logic in this function must, however, be atomic
with respect to changes to the callbacks_enabled field and destruction of
the pcon itself.  To protect against such races, a spinlock is used.

Signed-off-by: Martin Carroll <martin.carroll@alcatel-lucent.com>