vcrtcm: narrow range of spin lock
The alloc spin lock was being held during the calls to vcrtcm_get_pim
and vcrtcm_get_pcon, both of which lock a mutex.  Attempting to lock
a mutex while holding a spin lock is against the rules.  This commit
narrows the range of the spin lock, while still preserving the
protection of the critical section.

Signed-off-by: Martin Carroll <martin.carroll@alcatel-lucent.com>