vcrtcm: define locking variants of gpu-side functions
The pcon locking scheme is being rearchitected.  In particular,
the gpu-side api functions are going to be changed to *not* take
the pcon mutex, just as the pim-side api functions do not take the
pcon mutex.  Locking variants will also be provided.  It will be
the caller's responsibility, on both sides of the api, to ensure
that the pcon is locked before performing any operation on it.
The following sequence of commits effects the change on the gpu
side without breaking the build or the functioning of the system.
This commit simply defines the new locking variants of the gpu-
side functions.

Signed-off-by: Martin Carroll <martin.carroll@alcatel-lucent.com>