vcrtcm: change signatures of gpu callback functions
The gpu side of the vcrtcm api is changing to eliminate the need
(and ability) of the gpu driver to directly manipulate vcrtcm's
pcon structure.  The change will be implemented over a series
of commits.  In this first commit, the signatures of the gpu callback
functions are changed.  Note that this commit breaks the build;
the following commit will fix it.

Signed-off-by: Martin Carroll <martin.carroll@alcatel-lucent.com>