vcrtcm: change signatures of vcrtcm_g_attach
The gpu side of the vcrtcm api is changing to eliminate the need
(and ability) of the gpu driver to directly manipulate vcrtcm's
pcon structure.  The change will be implemented over a series
of commits.  In this commit, the signatures of vcrtcm_g_attach
and vcrtcm_g_attach_l are changed.  Note that this commit breaks
the build; the following commit will fix it.

Signed-off-by: Martin Carroll <martin.carroll@alcatel-lucent.com>