vcrtcm: change destroy callback to not hold pcon lock
The semantics of the destroy callback are changing.  As of this
commit, the destroy() callback is called with the pcon *not*
locked.   Not locking the pcon enables the PIM to wait_sync()
for other threads to finish all pcon-related operations without
deadlocking with the thread that is receiving this callback.
Note that this commit introduces more mutex violations (and
hence race conditions) into the system, because the pims are
currently implemented to assume that the pcon is locked when
destroy() is called.  These mutex violations will be fixed
in later commits.

Signed-off-by: Martin Carroll <martin.carroll@alcatel-lucent.com>