drm/radeon/kms: wait on correct fence in wait_fb callback
wait_fb callback is used to wait for push operation to finish.
We have the right fences to wait on (it's in the radeon_crtc
structure) so there is no need to wait for all existing
fences.

This patch also plugs a memory leak that was caused by waiting
in frame N on a fence from frame N-1. On detach, there was a
possiblity that the very last fence was never unreferenced.
With this patch, if there is an outstanding last fence, we
will unreference it on detach.

Signed-off-by: Ilija Hadzic <ihadzic@research.bell-labs.com>