v4l2pim: call wait_fb in detach, not destroy
Calling wait_fb function in destroy_pcon has little value
because by that time all transmission is gone, all push buffers
freed and and PCON is detached. So there is nothing to wait on.
The correct place to wait is the detach function: after we
have stopped the vblank thread and no new pushes are being
scheduled, we still have to wait for the last outstanding
one (which could be in the GPU's queue) to finish.

Signed-off-by: Ilija Hadzic <ihadzic@research.bell-labs.com>