vcrtcm: change pcon-table mutex to spin lock
In userspace, a mutex is virtually always preferable to a spin lock.
But in the kernel, if a spin lock is feasible (in particular, if use
of a spin lock would not appreciably affect system latency and would
not violate atomicity requirements), then use of a spin lock is
preferable over a mutex, because use of a mutex in a function precludes
that function from being called in atomic context.

In this particular case, the pcon table lookup routine is called
indirectly from the gpu driver's vblank callback, whose call chain
goes through the drm function drm_handle_vblank(), which takes a spin
lock, thereby putting the thread in atomic context.

Note that the conversion of the mutex to a spin lock necessitated some
rearrangement of the logic, to ensure that nothing nonatomic is called
while the spin lock is held.

Signed-off-by: Martin Carroll <martin.carroll@alcatel-lucent.com>