vcrtcm: change names of pcon_funcs structs
The vcrtcm_gpu_funcs struct was misnamed.  These functions are
not per-gpu, they are per-pcon, as evidenced by the fact that
the gpu driver typically specifies one set of functions for real
crts and a different set of functions for virtual crtcs.  This
commit also changes the name of the pcon_funcs field for clarity.
Note that this commit breaks the build; the following sequence
of commits will fix it.

Signed-off-by: Martin Carroll <martin.carroll@alcatel-lucent.com>