vcrtcm: add commentary to api functions
The commentary specifies, among other things, the atomicity and
pcon-locking guarantees and requirements of all the api functions.
Note that the stated properties of vcrtcm_p_emulate_vblank() do
not, as of this commit, match the implementation of that function;
this discrepancy will be fixed in the following sequence of commits.

Signed-off-by: Martin Carroll <martin.carroll@alcatel-lucent.com>