vcrtcm: move spin lock to pcon table entry
The page_flip logic was not protected against all pcon-destruction
races; in particular, it was possible for the destruction logic to
free the pcon struct out from under the page_flip logic.  To prevent
all page_flip-destruction races, we do for the spin lock what we
already did a while back for the pcon mutex: We associate it with
the pcon *id* intead of the pcon itself.  We must also remember to
spinlock the freeing of the pcon, because the page_flip logic needs
to look at the fields of that struct.

Signed-off-by: Martin Carroll <martin.carroll@alcatel-lucent.com>