vcrtcm: fix vcrtcm_p_emulate_vblank locking
The current implementation of vcrtcm_p_emulate_vblank() requires
the caller to hold the pcon's mutex.  This is wrong, because some
pims need to call that function in an ISR, where it is illegal to
take a mutex.  This commit fixes the implementation, and the api
documentation, to allow the function to be called without the
mutex being held.  The fixed implementation uses the pcon's spin
lock to protect against races.

Signed-off-by: Martin Carroll <martin.carroll@alcatel-lucent.com>