v4l2pim: change shadowbuf lock from mutex to spin lock
The shadowbuf is accessed within v4l2pim_fillbuff(), which is
called from v4l2pim_thread_tick() while the latter function is
holding a spin lock.  It is against the rules to take a mutex
while in atomic context.  This commit changes the shadowbuf
lock from a mutex to a spin lock.  Note that some of the logic
(specifically, the shadowbuf-freeing logic) had to be rearranged
so that the actually freeing, which is nonatomic, is not done
while the spin lock is held.

Signed-off-by: Martin Carroll <martin.carroll@alcatel-lucent.com>