vcrtcm: fix vblank worker cancellation logic
The commit fixes bugs related to the cancellation of the vblank
worker.  In particular, the implementation of the detach ioctl
was forgetting to cancel the vblank worker thread.  This bug
resulted in calls to the pim's vblank callback when the pcon is
detached.  Such calls violate the vcrtcm api and result in undefined
behavior.  This commit also adds an additional check to the vblank
worker itself to make sure that the pcon is attached before calling
the callback.

Signed-off-by: Martin Carroll <martin.carroll@alcatel-lucent.com>