vcrtcm: change lock requirement of vcrtcm_g_get_fb_status
The function vcrtcm_g_get_fb_status() can be called in a variety
of contexts, in some of which the pcon's mutex is locked, in others
of which the mutex is not locked.  To make it easier to implement
user code, this commit relaxes the lock requirement of this function
from "mutex must be locked" to "mutex need not be locked."  The
modified implementation uses the pcon's spin lock to protect against
race conditions.  The implementation is made rather more complex
by the fact that the function can also be called in contexts in
which the spin lock is already held.

Signed-off-by: Martin Carroll <martin.carroll@alcatel-lucent.com>